Make it possible for ESR to release environments after ETR ends
ESR will now start ETR instances as environments become available
and it will release environments as ETR instances finishes.
There was also some general cleanup made to make it easier to
understand what is going on.
This cleanup was necessary since I could not understand the
resulting code when I just patched in this support.